feat: log controller ingestion test, v2_pipeline sources table migration
