+++ lib/logflare/backends.ex
@@ -5,7 +5,8 @@ defmodule Logflare.Backends do
     SourceDispatcher,
     SourceRegistry,
     SourceBackendRegistry,
-    SourceSup
+    SourceSup,
+    SourcesSup
   }
 
   alias Logflare.Buffers.MemoryBuffer
@@ -28,7 +29,9 @@ defmodule Logflare.Backends do
   Adds log events to the source event buffer.
   The ingestion pipeline then pulls from the buffer and dispatches log events to the correct backends.
   """
-  def ingest_log_events(log_events, source) do
+  @type log_param :: map()
+  @spec ingest_logs(list(log_param()), Source.t()) :: :ok
+  def ingest_logs(log_events, source) do
     via = via_source(source, :buffer)
     MemoryBuffer.add_many(via, log_events)
   end
@@ -36,11 +39,12 @@ defmodule Logflare.Backends do
   @doc """
   Dispatch log events to a given source backend.
   It requires the source supervisor and registry to be running.
+  For internal use only, should not be called outside of the `Logflare` namespace.
   """
   def dispatch_ingest(log_events, source) do
     Registry.dispatch(SourceDispatcher, source.id, fn entries ->
       for {pid, {adaptor_module, :ingest}} <- entries do
-      # TODO: spawn tasks to do this concurrently
+        # TODO: spawn tasks to do this concurrently
         apply(adaptor_module, :ingest, [pid, log_events])
       end
     end)
@@ -50,12 +54,45 @@ defmodule Logflare.Backends do
 
   @doc """
   Registeres a unique source-related process on the source registry. Unique.
+  For internal use only, should not be called outside of the `Logflare` namespace.
   """
-  @spec via_source(Source.t(), atom()) :: tuple()
-  def via_source(%Source{id: id}, process_type),
-    do: {:via, Registry, {SourceRegistry, {id, process_type}}}
+  @spec via_source(Source.t(), term()) :: tuple()
+  def via_source(%Source{id: id}, process_id),
+    do: {:via, Registry, {SourceRegistry, {id, process_id}}}
 
   @spec via_source_backend(SourceBackend.t()) :: tuple()
   def via_source_backend(%SourceBackend{id: id}),
     do: {:via, Registry, {SourceBackendRegistry, id}}
+
+  @spec source_sup_started?(Source.t()) :: boolean()
+  def source_sup_started?(%Source{id: id}),
+    do: Registry.lookup(SourceRegistry, {id, SourceSup}) != []
+
+  @spec start_source_sup(Source.t()) :: :ok | {:error, :already_started}
+  def start_source_sup(%Source{} = source) do
+    with {:ok, _pid} <- DynamicSupervisor.start_child(SourcesSup, {SourceSup, source}) do
+      :ok
+    else
+      {:error, {:already_started = reason, _pid}} -> {:error, reason}
+    end
+  end
+
+  @spec stop_source_sup(Source.t()) :: :ok | {:error, :not_started}
+  def stop_source_sup(%Source{} = source) do
+    with [{pid, _}] <- Registry.lookup(SourceRegistry, {source.id, SourceSup}),
+         :ok <- DynamicSupervisor.terminate_child(SourcesSup, pid) do
+      :ok
+    else
+      _ -> {:error, :not_started}
+    end
+  end
+
+  @spec restart_source_sup(Source.t()) ::
+          :ok | {:error, :already_started} | {:error, :not_started}
+  def restart_source_sup(%Source{} = source) do
+    with :ok <- stop_source_sup(source),
+         :ok <- start_source_sup(source) do
+      :ok
+    end
+  end
 end
