+++ test/logflare/backends_test.exs
@@ -1,7 +1,8 @@
 defmodule Logflare.BackendsTest do
   @moduledoc false
   use Logflare.DataCase
-  alias Logflare.{LogEvent, Backends, Backends.SourceBackend, Backends.SourceSup}
+  alias Logflare.{Backends, Backends.SourceBackend, Backends.SourceSup}
+
   describe "backend management" do
     setup do
       user = insert(:user)
@@ -14,22 +15,62 @@ defmodule Logflare.BackendsTest do
     end
   end
 
-  describe "ingestion" do
-    setup :set_mimic_global
+  describe "SourceSup management" do
     setup do
       user = insert(:user)
       source = insert(:source, user_id: user.id)
-      insert(:source_backend, type: :webhook, source_id: source.id, config: %{url: "https://some-url.com"})
+      {:ok, source: source}
+    end
+
+    test "source_sup_started?/1", %{source: source} do
+      assert false == Backends.source_sup_started?(source)
+      start_supervised!({SourceSup, source})
+      :timer.sleep(400)
+      assert true == Backends.source_sup_started?(source)
+    end
+
+    test "start_source_sup/1, stop_source_sup/1, restart_source_sup/1", %{source: source} do
+      assert :ok = Backends.start_source_sup(source)
+      assert {:error, :already_started} = Backends.start_source_sup(source)
+
+      assert :ok = Backends.stop_source_sup(source)
+      assert {:error, :not_started} = Backends.stop_source_sup(source)
+
+      assert {:error, :not_started} = Backends.restart_source_sup(source)
+      assert :ok = Backends.start_source_sup(source)
+      assert :ok = Backends.restart_source_sup(source)
+    end
+  end
+
+  describe "ingestion" do
+    setup :set_mimic_global
+
+    setup do
+      user = insert(:user)
+      source = insert(:source, user_id: user.id)
+
+      insert(:source_backend,
+        type: :webhook,
+        source_id: source.id,
+        config: %{url: "https://some-url.com"}
+      )
+
       start_supervised!({SourceSup, source})
       {:ok, source: source}
     end
 
     test "backends receive dispatched log events", %{source: source} do
       Backends.Adaptor.WebhookAdaptor
-      |> expect(:ingest, fn _, _ -> :ok end)
+      |> expect(:ingest, fn _pid, [event | _] ->
+        if match?(%_{}, event) do
+          :ok
+        else
+          raise "Not a log event struct!"
+        end
+      end)
 
-      log_event = %LogEvent{}
-      assert :ok = Backends.ingest_log_events([log_event, log_event], source)
+      log_event = %{some: "event"}
+      assert :ok = Backends.ingest_logs([log_event, log_event], source)
       :timer.sleep(1500)
     end
   end
