+++ test/test_helper.exs
@@ -14,6 +14,7 @@ Mimic.copy(Logflare.SQL)
 Mimic.copy(Stripe.SubscriptionItem.Usage)
 Mimic.copy(Logflare.Backends.Adaptor.WebhookAdaptor)
 Mimic.copy(Logflare.Backends.Adaptor.WebhookAdaptor.Client)
+Mimic.copy(LogflareWeb.Plugs.RateLimiter)
 Mimic.copy(GoogleApi.BigQuery.V2.Api.Jobs)
 Mimic.copy(Goth)
 {:ok, _} = Application.ensure_all_started(:ex_machina)
